requirements.txt
# Spark
pyspark==v2.3.2
pyspark-stubs==2.3.0.post5

# Google Libraries
google-cloud-bigquery==1.5.0

